Use fixed width Size materialization

+++ crates/tune_shape/src/materialize.rs
@@ -34,7 +34,7 @@ pub fn can_materialize(lit: &LiteralFact, target: &Shape) -> bool {
             numeric_value(text).is_some_and(|n| n <= i64::MAX as u128)
         }
         (LiteralFact::Numeric { text }, Shape::Size) => {
-            numeric_value(text).is_some_and(|n| n <= usize::MAX as u128)
+            numeric_value(text).is_some_and(|n| n <= u64::MAX as u128)
         }
         (LiteralFact::Numeric { text }, Shape::Float) => float_value(text).is_some(),
         (LiteralFact::String { .. }, Shape::String) => true,
